From Frida Schlaug:
Imageio for ios had a poorly implemented function for reading from streams (always returning null). In this patch I copied some code from the non-ios imageio file and adjusted it to work on ios. (There is some refactoring to be done, but I don't have the means to test it anymore, so I'll keep my hands off it for now.)